apply dark mode to all forms automatically Based on https://github.com/kbilsted/NotepadPlusPlusPluginPack.Net/pull/104 Rdipardo did all the hard work here, this is just a refinement.
This makes it easy for plugin makers to apply dark mode to all forms.
Rather than having to individually select themes
to apply for each control, sensible defaults are applied based on
each control's type.
This PR also allows the color changes to propagate recursively
to a form's child forms, if it has any.